requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.3.0

# Code quality
flake8==6.0.0
black==23.3.0
isort==5.12.0

# Documentation
sphinx==7.0.1
sphinx-rtd-theme==1.2.0

# Debugging
ipdb==0.13.13
django-debug-toolbar==4.1.0
